Move the queued_blocks tests into their own module
