+++ zebra-state/src/service/queued_blocks/tests.rs
@@ -1,0 +1,3 @@
+//! Tests for block queues.
+
+mod vectors;
